further refactoring, bugfix
moved file handling functions into a separate file
removed debug outputs
bugfix: the select meme button was removed by accident
